+++ app/[locale]/(farmer)/dashboard/@recent/page.tsx
