+++ app/[locale]/(farmer)/dashboard/@statistics/loading.tsx
