+++ app/[locale]/(farmer)/dashboard/@statistics/page.tsx
