+++ app/[locale]/(farmer)/journal/page.tsx
