+++ app/[locale]/(farmer)/layout.tsx
