+++ app/[locale]/(technician)/checklist/[id]/layout.tsx
