+++ app/[locale]/(technician)/checklist/[id]/page.tsx
