+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@profile/loading.tsx
