+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@profile/page.tsx
