+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@summary/error.tsx
