+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@summary/loading.tsx
