+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@summary/page.tsx
