+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@visits/error.tsx
