+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@visits/loading.tsx
