+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/@visits/page.tsx
