+++ app/[locale]/(technician)/farmer/[id]/(dashboard)/layout.tsx
