+++ app/[locale]/(technician)/farmer/[id]/(others)/visits/layout.tsx
