+++ app/[locale]/(technician)/farmer/[id]/(others)/visits/loading.tsx
