+++ app/[locale]/(technician)/farmer/[id]/(others)/visits/page.tsx
