+++ app/[locale]/(technician)/layout.tsx
