+++ app/[locale]/(technician)/technician/@recent/loading.tsx
@@ -1,4 +1,4 @@
-import FarmerListItemSkeleton from "../../../../components/farmers/FarmerListItemSkeleton"
+import FarmerListItemSkeleton from "../../../../../components/farmers/FarmerListItemSkeleton"
 
 export default function Loading() {
   return Array.from({ length: 5 }).map((_, i) => <FarmerListItemSkeleton key={i} />)
