+++ app/[locale]/(technician)/technician/@statistics/error.tsx
