+++ app/[locale]/(technician)/technician/@statistics/loading.tsx
