+++ app/[locale]/(technician)/technician/@statistics/page.tsx
