+++ app/[locale]/(technician)/technician/layout.tsx
