+++ app/[locale]/(technician)/visit/[id]/layout.tsx
