+++ app/[locale]/(technician)/visit/[id]/loading.tsx
