+++ app/[locale]/(technician)/visit/[id]/page.tsx
