+++ app/[locale]/logout/route.ts
