+++ i18n.ts
@@ -1,0 +1,5 @@
+import {getRequestConfig} from 'next-intl/server';
+ 
+export default getRequestConfig(async ({locale}) => ({
+  messages: (await import(`./messages/${locale}.json`)).default
+}));
