+++ package.json
@@ -22,6 +22,7 @@
     "jsonapi-fractal": "^2.3.1",
     "jwt-decode": "^3.1.2",
     "next": "13.4.7",
+    "next-intl": "^3.0.0-beta.6",
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
